+++ spaceinvaders/src/model/aliens/ships/queue-aliens/EnemyGroup.cs
@@ -1,7 +1,7 @@
 using Microsoft.Xna.Framework.Graphics;
 
 public interface IEnemyGroup : IEnemyEntity {
-    public void IncreaseX(int x);
+    public void IncreaseX(float x);
     public void InvertDirection();
     public Texture2D GetTexture();
     public void Fall();
