+++ spaceinvaders/src/model/aliens/ships/queue-aliens/ShooterEnemy.cs
@@ -82,8 +82,8 @@ public class ShooterEnemy : IEnemyGroup
         directionRight = !directionRight;
     }
 
-    public void IncreaseX(int value) {
-        int valueModified = directionRight ? value : -value;
+    public void IncreaseX(float value) {
+        float valueModified = directionRight ? value : -value;
         Bounds.Position += new Vector2(valueModified, 0);
     }
